--- packages/client-p2p/src/server/onPeerMessage.js
@@ -1,20 +1,0 @@
-// Copyright 2017-2018 Jaco Greeff
-// This software may be modified and distributed under the terms
-// of the ISC license. See the LICENSE file for details.
-// @flow
-
-import type { MessageInterface, PeerInterface } from '../types';
-import type { P2pState } from './types';
-
-type Message = {
-  peer: PeerInterface,
-  message: MessageInterface
-};
-
-const handleMessage = require('../handler');
-
-module.exports = function onPeerMessage (self: P2pState): void {
-  self.peers.on('message', ({ peer, message }: Message): void => {
-    handleMessage(self, peer, message);
-  });
-};
