+++ packages/client-p2p/src/server/onPeerMessage/blockAnnounce.js
@@ -3,15 +3,15 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { MessageInterface, PeerInterface } from '../types';
-import type { BlockAnnounceMessage } from '../message/types';
-import type { P2pState } from '../server/types';
+import type { BlockAnnounceMessage } from '../../message/types';
+import type { P2pState } from '../../server/types';
+import type { MessageInterface, PeerInterface } from '../../types';
 
 const encodeHeader = require('@polkadot/primitives-codec/header/encode');
 const blake2Asu8a256 = require('@polkadot/util-crypto/blake2/asU8a256');
 
-const message = require('../message/blockAnnounce');
-const requestsBlocks = require('../sync/requestBlocks');
+const message = require('../../message/blockAnnounce');
+const requestsBlocks = require('../../sync/requestBlocks');
 
 module.exports = function handleBlockAnnounce (self: P2pState, peer: PeerInterface, message: MessageInterface): void {
   self.l.debug(() => [peer.shortId, 'BlockAnnounce', JSON.stringify(message.encode().message)]);
