+++ packages/client-p2p/src/server/onPeerMessage/blockResponse.js
@@ -3,12 +3,12 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { BlockResponseMessage } from '../message/types';
-import type { P2pState } from '../server/types';
-import type { MessageInterface, PeerInterface } from '../types';
+import type { BlockResponseMessage } from '../../message/types';
+import type { P2pState } from '../../server/types';
+import type { MessageInterface, PeerInterface } from '../../types';
 
-const message = require('../message/blockResponse');
-const queueBlocks = require('../sync/queueBlocks');
+const message = require('../../message/blockResponse');
+const queueBlocks = require('../../sync/queueBlocks');
 
 module.exports = function handleBlockResponse (self: P2pState, peer: PeerInterface, message: MessageInterface): void {
   self.l.debug(() => [peer.shortId, 'BlockResponse', JSON.stringify(message.encode().message)]);
