+++ packages/client-p2p/src/server/onPeerMessage/index.js
@@ -3,8 +3,13 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { MessageInterface, PeerInterface } from '../types';
-import type { P2pState } from '../server/types';
+import type { MessageInterface, PeerInterface } from '../../types';
+import type { P2pState } from '../types';
+
+type Message = {
+  peer: PeerInterface,
+  message: MessageInterface
+};
 
 type Handler = {
   (self: P2pState, peer: PeerInterface, message: MessageInterface): void,
@@ -20,18 +25,15 @@ const HANDLERS: Array<Handler> = [
   blockAnnounce, blockRequest, blockResponse, status
 ];
 
-module.exports = function handleMessage (self: P2pState, peer: PeerInterface, message: MessageInterface): void {
-  self.emitter.emit('message', {
-    peer,
-    message
+module.exports = function onPeerMessage (self: P2pState): void {
+  self.peers.on('message', ({ peer, message }: Message): void => {
+    const handler = HANDLERS.find((handler) => handler.TYPE === message.type);
+
+    if (!handler) {
+      self.l.error(`Unhandled message type=${message.type}`);
+      return;
+    }
+
+    handler(self, peer, message);
   });
-
-  const handler = HANDLERS.find((handler) => handler.TYPE === message.type);
-
-  if (!handler) {
-    self.l.error(`Unhandled message type=${message.type}`);
-    return;
-  }
-
-  return handler(self, peer, message);
 };
